Cleanup embedding code and command-line flags
Remove all flags that are always disabled (like the Transformer
layers on the decoder side) and set the dimension based on
the pretrained model size.
