+++ genienlp/cache_embeddings.py
@@ -29,9 +29,8 @@
 
 
 import logging
-from pprint import pformat
+from transformers import AutoModel, AutoTokenizer
 
-from .data_utils.embeddings import load_embeddings
 from .util import set_seed
 
 logger = logging.getLogger(__name__)
@@ -44,7 +43,8 @@ def parse_argv(parser):
 
 
 def main(args):
-    logger.info(f'Arguments:\n{pformat(vars(args))}')
-
     set_seed(args)
-    load_embeddings(args.destdir, args.embeddings, '', '', cache_only=True)
+
+    for embedding in args.embeddings.split('+'):
+        AutoTokenizer.from_pretrained(embedding, cache_dir=args.destdir)
+        AutoModel.from_pretrained(embedding, cache_dir=args.destdir)
