+++ genienlp/models/__init__.py
@@ -27,4 +27,5 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from .general_seq2seq import Seq2Seq, Bart
+from .bert_lstm import BertLSTM
+from .bart import Bart
