+++ genienlp/models/common.py
@@ -40,11 +40,6 @@ from torch.nn.utils.rnn import pack_padded_sequence as pack
 from torch.nn.utils.rnn import pad_packed_sequence as unpack
 
 
-class EmbeddingOutput(NamedTuple):
-    all_layers: List[torch.Tensor]
-    last_layer: torch.Tensor
-
-
 INF = 1e10
 EPSILON = 1e-10
 
@@ -407,33 +402,13 @@ class CombinedEmbedding(nn.Module):
 
 
     def _combine_embeddings(self, embeddings):
-        if len(embeddings) == 1:
-            all_layers = embeddings[0].all_layers
-            last_layer = embeddings[0].last_layer
-            if self.project:
-                last_layer = self.projection(last_layer)
-            return EmbeddingOutput(all_layers=all_layers, last_layer=last_layer)
-
-        all_layers = None
-        last_layer = []
-        for emb in embeddings:
-            if all_layers is None:
-                all_layers = [[layer] for layer in emb.all_layers]
-            elif len(all_layers) != len(emb.all_layers):
-                raise ValueError('Cannot combine embeddings that use different numbers of layers')
-            else:
-                for layer_list, layer in zip(all_layers, emb.all_layers):
-                    layer_list.append(layer)
-            last_layer.append(emb.last_layer)
-
-        all_layers = [torch.cat(layer, dim=2) for layer in all_layers]
-        last_layer = torch.cat(last_layer, dim=2)
+        emb = torch.cat(embeddings, dim=2)
         if self.project:
-            last_layer = self.projection(last_layer)
-        return EmbeddingOutput(all_layers=all_layers, last_layer=last_layer)
+            emb = self.projection(emb)
+        return emb
 
     def forward(self, x, padding=None):
-        embedded: List[EmbeddingOutput] = []
+        embedded: List[torch.Tensor] = []
         if self.pretrained_embeddings is not None:
             embedded += [emb(x, padding=padding) for emb in self.pretrained_embeddings]
 
@@ -442,7 +417,7 @@ class CombinedEmbedding(nn.Module):
             valid_x = torch.lt(x, trained_vocabulary_size)
             masked_x = torch.where(valid_x, x, torch.zeros_like(x))
             output = self.trained_embeddings(masked_x)
-            embedded.append(EmbeddingOutput(all_layers=[output], last_layer=output))
+            embedded.append(output)
 
         return self._combine_embeddings(embedded)
