+++ genienlp/tasks/generic_task.py
@@ -141,13 +141,6 @@ class BaseSummarizationTask(BaseTask):
     def metrics(self):
         return ['avg_rouge', 'rouge1', 'rouge2', 'rougeL', 'em', 'nem', 'nf1']
 
-    def preprocess_example(self, ex, train=False, max_context_length=None):
-        # Filter examples with a dummy summary
-        if train and 'This page includes the show' in ex.answer:
-            return None
-
-        return ex._replace(context=ex.context[:max_context_length])
-
 
 @register_task('cnn')
 class CNN(BaseSummarizationTask):
@@ -179,9 +172,6 @@ class SST(BaseTask):
 
 @register_task('imdb')
 class IMDB(BaseTask):
-    def preprocess_example(self, ex, train=False, max_context_length=None):
-        return ex._replace(context=ex.context[:max_context_length])
-
     def get_splits(self, root, **kwargs):
         kwargs['validation'] = None
         return generic_dataset.IMDb.splits(root=root, **kwargs)
